+++ MarketScanner/Types/Option.cs
@@ -11,11 +11,11 @@ namespace MarketScanner.Types
         public DateTime QuoteDate { get; set; }
         public double Strike { get; set; }
         public string Last { get; set; }
-        public double Bid { get; set; }
-        public double Ask { get; set; }
+        public decimal Bid { get; set; }
+        public decimal Ask { get; set; }
         public string Volume { get; set; }
         public string OpenInterest { get; set; }
         public double Delta { get; set; }
-        public double StockPrice { get; set; }
+        public decimal StockPrice { get; set; }
     }
 }
